Organize transcription audit into submodules
