+++ scinoephile/analysis/audit/transcription/__init__.py
@@ -1,10 +1,8 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Subtitle audit reporting and analysis.
+"""Portable multi-source transcription alignment audits.
 
 Package hierarchy (modules may import from any above):
-* utils
-* aligned_diff / delineation / ocr_fusion / punctuation / review
-  / transcription / translation
-* dual_review / gap_translation / guided_review / guided_translation
+* rendering
+* report
 """
